Wrap toaster exports in namespace to streamline usages
I got so fucking sick of not being able to let the IDE generate the
import, so the namespace will ensure that there's a recognizable name we
can start typing for autocompletion ("toaster"), and the single import
will be enough to let us keep using the toaster like we have been, e.g.
`toaster.success("Success.")`.